% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_fragilite}
\alias{res_fragilite}
\title{liste contenant les résultats de la méthode fragilité avec les paramètres par défaut}
\format{
A list
}
\usage{
data(res_fragilite)
}
\description{
\itemize{
  \item table_score_EG_clean une ligne par établissement et le résultat pour chaque indicateur/famille/type etc.
  \item table_score_clean Une ligne par niveau par indicateur et le score associé
  \item list_data_fiche_etab une liste contenant une fiche de score par établissement
}
}
\keyword{datasets}
