% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab_context}
\alias{tab_context}
\title{table contenant les infos de contextualisation des établissements}
\format{
A data frame
}
\usage{
data(tab_context)
}
\description{
table contenant les infos de contextualisation des établissements
}
\keyword{datasets}
