% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab_dpt_region}
\alias{tab_dpt_region}
\title{table contenant les départements et régions de France}
\format{
A data frame
}
\usage{
data(tab_dpt_region)
}
\description{
table contenant les départements et régions de France
}
\keyword{datasets}
