% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab_indic}
\alias{tab_indic}
\title{table contenant les indicateurs calculés pour chaque établissement}
\format{
A data frame
}
\usage{
data(tab_indic)
}
\description{
table contenant les indicateurs calculés pour chaque établissement
}
\keyword{datasets}
